refactoring of multi-steps form
