+++ src/components/pages/Booking/Booking.jsx
@@ -1,47 +1,12 @@
 import { useState } from "react";
+// Components
 import PersonalDetails from "./PersonalDetails";
 import TableDetails from "./TableDetails";
 import Confirmation from "./Confirmation";
-import { Formik, Form } from "formik";
-import * as Yup from "yup";
-import SuccessPopup from "./SuccessPopup";
-
-const initialValues = {
-  date: new Date().toJSON().slice(0, 10),
-  time: new Date().toLocaleTimeString(),
-  dinners: 2,
-  occasion: "none",
-  seatingOptions: "Indoor",
-  specialRequest: "",
-  firstName: "",
-  lastName: "",
-  email: "",
-  password: "",
-  cardNumber: "",
-  nameOnCard: "",
-  expDate: "",
-  ccv: "",
-};
-
-const validationSchema = Yup.object().shape({
-  date: Yup.date().required("Required"),
-  time: Yup.string().required("Required"),
-  dinners: Yup.number().required("Required"),
-  firstName: Yup.string().required("Required"),
-  lastName: Yup.string().required("Required"),
-  email: Yup.string().required("Required"),
-  password: Yup.string().required("Required"),
-  cardNumber: Yup.string().required("Required"),
-  nameOnCard: Yup.string().required("Required"),
-  expDate: Yup.string().required("Required"),
-  ccv: Yup.string().required("Required"),
-});
 
 export default function Booking() {
+  const [formData, setFormData] = useState({});
   const [step, setStep] = useState(1);
-  const [isSubmitted, setSubmitted] = useState(false);
-
-  const onSubmit = (values) => setSubmitted(true);
 
   const updateStep = (action = "next") => {
     switch (action) {
@@ -58,27 +23,23 @@ export default function Booking() {
 
   return (
     <>
-      <div className={`popup ${!isSubmitted ? "hidden" : ""}`}>
-        <SuccessPopup />
-      </div>
-      <Formik
-        initialValues={initialValues}
-        validationSchema={validationSchema}
-        onSubmit={onSubmit}>
-        {(formik) => (
-          <Form>
-            {step === 1 ? (
-              <TableDetails formik={formik} updateStep={updateStep} />
-            ) : step === 2 ? (
-              <PersonalDetails formik={formik} updateStep={updateStep} />
-            ) : step === 3 ? (
-              <Confirmation formik={formik} updateStep={updateStep} />
-            ) : (
-              ""
-            )}
-          </Form>
-        )}
-      </Formik>
+      {step === 1 ? (
+        <TableDetails updateStep={updateStep} setFormData={setFormData} />
+      ) : step === 2 ? (
+        <PersonalDetails
+          updateStep={updateStep}
+          formData={formData}
+          setFormData={setFormData}
+        />
+      ) : step === 3 ? (
+        <Confirmation
+          updateStep={updateStep}
+          formData={formData}
+          setFormData={setFormData}
+        />
+      ) : (
+        ""
+      )}
     </>
   );
 }
